Moved GUI log in ft.
